# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: 'SAKA Marketplace API'
  version: 1.0.0
  description: |-
    Multi-vertical marketplace API for Tanzania.

    **Errors.** Every failure returns the same envelope:

        { "error": { "code": "...", "message": "...", "details": {...}, "request_id": "..." } }

    Clients should switch on `error.code`, never on `message`. Validation failures (422) additionally carry Laravel's `errors` map.

    **Correlation.** Every response carries `X-Request-Id`; an inbound value is honoured when it matches `^[A-Za-z0-9_-]{8,64}$`.

    **Authentication.** Bearer tokens issued by `/auth/login`. Tokens are unscoped; authorization is enforced by server-side policies.

    **Not found vs forbidden.** Resources the caller may not see return 404, never 403, so existence is not disclosed.
  contact:
    name: 'SAKA Engineering'
  license:
    name: Proprietary
servers:
  -
    url: 'https://api.saka.co.tz/api/v1'
    description: Production
  -
    url: 'http://localhost:8000/api/v1'
    description: Local
...
paths:
...
  '/admin/taxonomy/{type}':
    post:
      operationId: api_v1_admin_taxonomy_store
      summary: 'Create an amenity or facility'
      tags:
        - Administration
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Taxonomy'
        401:
          $ref: '#/components/responses/Unauthenticated'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        422:
          $ref: '#/components/responses/ValidationFailed'
        429:
          $ref: '#/components/responses/RateLimited'
        500:
          $ref: '#/components/responses/ServerError'
      description: |-
        Requires `amenity.manage`. Amenities and facilities share one shape and one set of routes, discriminated by `{type}`.

        Rate limit: `admin`.
      parameters:
        -
          name: type
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreTaxonomyTermRequest'
      security:
        -
          bearerAuth: []
  '/admin/taxonomy/{type}/{slug}':
    patch:
      operationId: api_v1_admin_taxonomy_update
      summary: 'Update an amenity or facility'
      tags:
        - Administration
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Taxonomy'
        401:
          $ref: '#/components/responses/Unauthenticated'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        422:
          $ref: '#/components/responses/ValidationFailed'
        429:
          $ref: '#/components/responses/RateLimited'
        500:
          $ref: '#/components/responses/ServerError'
      description: |-
        Requires `amenity.manage`. The slug is immutable — it is a public filter value.

        Rate limit: `admin`.
      parameters:
        -
          name: type
          in: path
          required: true
          schema:
            type: string
        -
          name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreTaxonomyTermRequest'
      security:
        -
          bearerAuth: []
    delete:
      operationId: api_v1_admin_taxonomy_destroy
      summary: 'Delete an amenity or facility'
      tags:
        - Administration
      responses:
        200:
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Message'
        401:
          $ref: '#/components/responses/Unauthenticated'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        409:
          $ref: '#/components/responses/Conflict'
        429:
          $ref: '#/components/responses/RateLimited'
        500:
          $ref: '#/components/responses/ServerError'
      description: |-
        Requires `amenity.manage`. Returns 409 when listings still reference the term — deactivate it instead.

        Rate limit: `admin`.
      parameters:
        -
          name: type
          in: path
          required: true
          schema:
            type: string
        -
          name: slug
          in: path
          required: true
          schema:
            type: string
      security:
        -
          bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 'Token from `/auth/login`, `/auth/register` or `/auth/oauth/google`.'
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            request_id:
              type: string
          required:
            - code
            - message
      required:
        - error
    ValidationError:
      allOf:
        -
          $ref: '#/components/schemas/Error'
        -
          type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
    Message:
      type: object
      properties:
        message:
          type: string
...
    Taxonomy:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        icon:
          type:
            - string
            - 'null'
...
    StoreTaxonomyTermRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        slug:
          type: string
          maxLength: 120
          nullable: true
          description: 'Create only; immutable afterwards.'
        icon:
          type: string
          maxLength: 60
          nullable: true
        category_id:
          type: integer
          nullable: true
          description: 'Scopes the term to one vertical.'
        position:
          type: integer
          minimum: 0
          maximum: 65535
        is_active:
          type: boolean
...
  responses:
    Unauthenticated:
      description: 'Missing or invalid token (`UNAUTHENTICATED`).'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 'Authenticated but not permitted (`FORBIDDEN`).'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 'Not found, or not visible to this caller (`NOT_FOUND`). Returned instead of 403 so existence is never disclosed.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: 'Conflicts with current state (`CONFLICT`, `PHONE_ALREADY_REGISTERED`, ...).'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ValidationFailed:
      description: 'Validation failed (`VALIDATION_FAILED`). Carries an `errors` map.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    RateLimited:
      description: 'Rate limit exceeded (`RATE_LIMITED`). See `Retry-After`.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: 'Unexpected error (`SERVER_ERROR`). No internal detail is disclosed.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'